+++ desktop/app/src/main/kotlin/com/abdownloadmanager/desktop/ui/Ui.kt
@@ -1,6 +1,12 @@
 package com.abdownloadmanager.desktop.ui
 
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.window.ApplicationScope
 import androidx.compose.ui.window.application
 import com.abdownloadmanager.desktop.AppArguments
@@ -21,6 +27,7 @@ import com.abdownloadmanager.desktop.pages.filehash.FileChecksumWindow
 import com.abdownloadmanager.desktop.pages.home.HomeWindow
 import com.abdownloadmanager.desktop.pages.newQueue.NewQueueDialog
 import com.abdownloadmanager.desktop.pages.queue.QueuesWindow
+import com.abdownloadmanager.desktop.pages.settings.FontManager
 import com.abdownloadmanager.desktop.pages.settings.SettingWindow
 import com.abdownloadmanager.desktop.pages.settings.ThemeManager
 import com.abdownloadmanager.desktop.pages.singleDownloadPage.ShowDownloadDialogs
@@ -34,6 +41,7 @@ import com.abdownloadmanager.shared.ui.theme.ABDownloaderTheme
 import com.abdownloadmanager.shared.ui.widget.ProvideLanguageManager
 import com.abdownloadmanager.shared.ui.widget.ProvideNotificationManager
 import com.abdownloadmanager.shared.ui.widget.useNotification
+import com.abdownloadmanager.shared.utils.LocalUseRelativeDateTime
 import com.abdownloadmanager.shared.utils.ProvideSizeAndSpeedUnit
 import com.abdownloadmanager.shared.utils.mvi.HandleEffects
 import com.abdownloadmanager.shared.utils.ui.ProvideDebugInfo
@@ -41,6 +49,7 @@ import com.abdownloadmanager.shared.utils.ui.icon.MyIcons
 import ir.amirab.util.compose.action.buildMenu
 import ir.amirab.util.compose.asStringSource
 import ir.amirab.util.compose.localizationmanager.LanguageManager
+import ir.amirab.util.desktop.PlatformDockToggler
 import ir.amirab.util.desktop.mac.event.MacEventHandler
 import ir.amirab.util.desktop.systemtray.IComposeSystemTray
 import ir.amirab.util.platform.Platform
@@ -60,8 +69,10 @@ object Ui : KoinComponent {
     ) {
         val appComponent: AppComponent = get()
         val themeManager: ThemeManager = get()
+        val fontManager: FontManager = get()
         val languageManager: LanguageManager = get()
         themeManager.boot()
+        fontManager.boot()
         languageManager.boot()
         if (!appArguments.startSilent) {
             appComponent.openHome()
@@ -79,61 +90,98 @@ object Ui : KoinComponent {
             )
         }
         application {
-            val theme by themeManager.currentThemeColor.collectAsState()
-            ProvideDebugInfo(AppInfo.isInDebugMode()) {
-                ProvideLanguageManager(languageManager) {
-                    ProvideNotificationManager {
-                        ABDownloaderTheme(
-                            myColors = theme,
-                            uiScale = appComponent.uiScale.collectAsState().value
-                        ) {
-                            ProvideGlobalExceptionHandler(globalAppExceptionHandler) {
-                                ProvideSizeUnits(appComponent) {
-                                    HandleEffectsForApp(appComponent)
-                                    SystemTray(appComponent)
-                                    val showHomeSlot =
-                                        appComponent.showHomeSlot.collectAsState().value
-                                    showHomeSlot.child?.instance?.let {
-                                        HomeWindow(it, appComponent::closeHome)
-                                    }
-                                    val showSettingSlot =
-                                        appComponent.showSettingSlot.collectAsState().value
-                                    showSettingSlot.child?.instance?.let {
-                                        SettingWindow(it, appComponent::closeSettings)
-                                    }
-                                    val showQueuesSlot =
-                                        appComponent.showQueuesSlot.collectAsState().value
-                                    showQueuesSlot.child?.instance?.let {
-                                        QueuesWindow(it)
-                                    }
-                                    val batchDownloadSlot =
-                                        appComponent.batchDownloadSlot.collectAsState().value
-                                    batchDownloadSlot.child?.instance?.let {
-                                        BatchDownloadWindow(it)
-                                    }
-                                    val editDownloadSlot =
-                                        appComponent.editDownloadSlot.collectAsState().value
-                                    editDownloadSlot.child?.instance?.let {
-                                        EditDownloadWindow(it)
-                                    }
-                                    ShowAddDownloadDialogs(appComponent)
-                                    ShowDownloadDialogs(appComponent)
-                                    ShowCategoryDialogs(appComponent)
-                                    FileChecksumWindow(appComponent)
-                                    ShowUpdaterDialog(appComponent.updater)
-                                    ShowAboutDialog(appComponent)
-                                    NewQueueDialog(appComponent)
-                                    ShowMessageDialogs(appComponent)
-                                    ShowOpenSourceLibraries(appComponent)
-                                    ShowTranslators(appComponent)
-                                    ConfirmExit(appComponent)
-                                }
+            ProvideLocalProviders(
+                languageManager = languageManager,
+                appComponent = appComponent,
+                themeManager = themeManager,
+                fontManager = fontManager,
+                globalAppExceptionHandler = globalAppExceptionHandler,
+            ) {
+                HandleEffectsForApp(appComponent)
+                SystemTray(appComponent)
+                val showHomeSlot =
+                    appComponent.showHomeSlot.collectAsState().value
+                showHomeSlot.child?.instance?.let {
+                    HomeWindow(it, appComponent::closeHome)
+                }
+                val showSettingSlot =
+                    appComponent.showSettingSlot.collectAsState().value
+                showSettingSlot.child?.instance?.let {
+                    SettingWindow(it, appComponent::closeSettings)
+                }
+                val showQueuesSlot =
+                    appComponent.showQueuesSlot.collectAsState().value
+                showQueuesSlot.child?.instance?.let {
+                    QueuesWindow(it)
+                }
+                val batchDownloadSlot =
+                    appComponent.batchDownloadSlot.collectAsState().value
+                batchDownloadSlot.child?.instance?.let {
+                    BatchDownloadWindow(it)
+                }
+                val editDownloadSlot =
+                    appComponent.editDownloadSlot.collectAsState().value
+                editDownloadSlot.child?.instance?.let {
+                    EditDownloadWindow(it)
+                }
+                ShowAddDownloadDialogs(appComponent)
+                ShowDownloadDialogs(appComponent)
+                ShowCategoryDialogs(appComponent)
+                FileChecksumWindow(appComponent)
+                ShowUpdaterDialog(appComponent.updater)
+                ShowAboutDialog(appComponent)
+                NewQueueDialog(appComponent)
+                ShowMessageDialogs(appComponent)
+                ShowOpenSourceLibraries(appComponent)
+                ShowTranslators(appComponent)
+                ConfirmExit(appComponent)
+            }
+        }
+    }
+}
+
+@Composable
+private fun ProvideLocalProviders(
+    languageManager: LanguageManager,
+    themeManager: ThemeManager,
+    fontManager: FontManager,
+    appComponent: AppComponent,
+    globalAppExceptionHandler: GlobalAppExceptionHandler,
+    content: @Composable () -> Unit
+) {
+    val theme by themeManager.currentThemeColor.collectAsState()
+    val fontFamily by fontManager.currentFontFamily.collectAsState()
+    ProvideDebugInfo(AppInfo.isInDebugMode()) {
+        ProvideLanguageManager(languageManager) {
+            ProvideCommonSettings(appComponent) {
+                ProvideNotificationManager {
+                    ABDownloaderTheme(
+                        myColors = theme,
+                        fontFamily = fontFamily,
+                        uiScale = appComponent.uiScale.collectAsState().value
+                    ) {
+                        ProvideGlobalExceptionHandler(globalAppExceptionHandler) {
+                            ProvideSizeUnits(appComponent) {
+                                content()
                             }
                         }
                     }
                 }
             }
         }
+    }
+}
+
+@Composable
+private fun ProvideCommonSettings(
+    appComponent: AppComponent,
+    content: @Composable () -> Unit,
+) {
+    val useNativeDateTime by appComponent.appSettings.useRelativeDateTime.collectAsState()
+    CompositionLocalProvider(
+        LocalUseRelativeDateTime provides useNativeDateTime
+    ) {
+        content()
     }
 }
 
@@ -172,6 +220,7 @@ private fun ApplicationScope.SystemTray(
 ) {
     val useSystemTray by component.useSystemTray.collectAsState()
     if (useSystemTray) {
+        LaunchedEffect(Unit) { PlatformDockToggler.hide() }
         IComposeSystemTray.Instance.ComposeSystemTray(
             icon = MyIcons.appIcon,
             onClick = showDownloadList,
@@ -184,5 +233,7 @@ private fun ApplicationScope.SystemTray(
                 }
             }
         )
+    } else {
+        LaunchedEffect(Unit) { PlatformDockToggler.show() }
     }
 }