+++ shared/app/src/main/kotlin/com/abdownloadmanager/shared/ui/theme/ABDownloaderTheme.kt
@@ -1,14 +1,13 @@
 package com.abdownloadmanager.shared.ui.theme
 
 import ir.amirab.util.compose.action.buildMenu
-import com.abdownloadmanager.shared.utils.darker
 import com.abdownloadmanager.shared.utils.div
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.*
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
@@ -42,54 +41,6 @@ fun MyColors.asMaterial2Colors(): Colors {
 */
 
 
-val darkColors = MyColors(
-    //used
-    primary = Color(0xFF4791BF),
-    onPrimary = Color.White,
-    secondary = Color(0xFFB85DFF),
-    onSecondary = Color.White,
-    background = Color(0xFF16161E),
-    onBackground = Color(0xFFAAAAAA),
-    onSurface = Color(0xFFAAAAAA),
-//        surface = Color(0xff191922),
-    surface = Color(0xFF22222A),
-    error = Color(0xffff5757),
-    onError = Color.White,
-    success = Color(0xff69BA5A),
-    onSuccess = Color.White,
-    warning = Color(0xFFffbe56),
-    onWarning = Color.White,
-    info = Color(0xFF2f77d4),
-    onInfo = Color.White,
-    isLight = false,
-    name = "Dark",
-    id = "dark",
-)
-val lightColors = MyColors(
-    primary = Color(0xFF4791BF),
-    primaryVariant = Color(0xFFAFCEFF).darker(0.5f),
-    onPrimary = Color.Black,
-    secondary = Color(0xFFB85DFF),
-    onSecondary = Color.White,
-//        primary = Color(0xFF3B82F6),
-    background = Color.White.darker(0.1f),
-    backgroundVariant = Color(0xFFfafafa),
-    onBackground = Color(0xFF353535),
-    onSurface = Color(0xFF353535),
-//        surface = Color(0xff191922),
-    surface = Color.White.darker(0.15f),
-    error = Color(0xffff5757),
-    onError = Color.White,
-    success = Color(0xff14a600),
-    onSuccess = Color.White,
-    warning = Color(0xFFffbe56),
-    onWarning = Color.White,
-    info = Color(0xFF2f77d4),
-    onInfo = Color.White,
-    isLight = true,
-    name = "Light",
-    id = "light",
-)
 
 private val textSizes = TextSizes(
     xs = 8.sp,
@@ -102,6 +53,7 @@ private val textSizes = TextSizes(
 @Composable
 fun ABDownloaderTheme(
     myColors: MyColors,
+    fontFamily: FontFamily? = null,
     uiScale: Float? = null,
     content: @Composable () -> Unit,
 ) {
@@ -121,6 +73,7 @@ fun ABDownloaderTheme(
             LocalTextStyle provides LocalTextStyle.current.copy(
                 lineHeight = TextUnit.Unspecified,
                 fontSize = textSizes.base,
+                fontFamily = fontFamily,
             ),
         ) {
             // it is overridden by [Window] Composable,